infra/tester: Add tutorial / battle tester
This python script takes two forms:

  tester.py tutorial NUM BINARY_PATH
  tester.py battle BINARY_PATH BINARY_PATH

The tutorial subcommand runs the bot against tutorial stages. There are
13 stages. The second argument takes '1' to '13'.

The battle subcommand takes two bots and let them fight each other.

Note that the tutorial stages requires the START request to be sent with
no argument. The script sets a special environment variable
(TUTORIAL_MODE) to let the bot know that.

In order to communicate with the API server, you need an API key. The
key should be set to the environment variable "ICFPC_API_KEY".